pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=2.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.10",
    "wtforms>=3.2.1",
    "flask-wtf>=1.2.2",
    "flask-mail>=0.10.0",
    "werkzeug>=2.0",
    "sqlalchemy>=2.0.41",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
    "markupsafe>=3.0.2",
    "msal>=1.32.3",
    "reportlab>=4.4.2",
    "xlsxwriter>=3.2.5",
    "pytz>=2025.2",
    "flask-migrate>=4.1.0",
    "stripe>=12.3.0",
    "pillow==12.1.1",
    "oauthlib>=3.2.2",
    "pyjwt>=2.10.1",
    "beautifulsoup4>=4.13.4",
    "cryptography==46.0.5",
    "python-dateutil>=2.9.0.post0",
    "flask-limiter>=3.12",
    "defusedcsv>=2.0.0",
    "defusedxml>=0.7.1",
    "flask-login>=0.6.3",
    "openpyxl>=3.1.5",
    "bleach>=6.3.0",
    "urllib3>=2.6.2",
]
